coach:     * removed empty results from athlete count     * tbs (training bests) instead of sbs (stagional bests)     * athlete results styling     * fixed (~) conversion from plan to scheduled-trainings     * !fixed insertion of result into current day instead of the correct one
athlete:
    * added change-results support
    * ScheduledTraining preview in main shows results
    * fixed corrupted result bug in main

+++ lib/ripetuta/ripetuta.dart
@@ -29,7 +29,7 @@ class Ripetuta {
   }
   Ripetuta.parse(Map raw) {
     template = raw['template'];
-    target = raw['target'];
+    target = raw['target'].toDouble();
     recupero = Recupero(raw['recupero']);
     ripetizioni = raw['times'];
     nextRecupero = Recupero(raw['recuperoNext']);
